fix: refactoring avatar component
