+++ src/components/Avatar/index.ts
@@ -1,2 +1,1 @@
 export * from './Avatar';
-export * from './AvatarGroup';
